wip: better utils in abd-clam
